Refactor code to create a directory to store server files
This makes it so that the server creates and checks a specific directory rather than having the client specify a specific path to a file. Now, the server will just need the name of the file within the directory it creates. This will also make it easier to list all files within the server later on. I also refactored the rename and delete function to ensure it works properly.